requirements.txt
discord.py
python-dotenv
docker
aiofiles
pyyaml
